web_scraper/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
pydantic==2.6.1
lxml==5.1.0
python-dateutil==2.8.2
